+++ src/main/java/com/deal4u/fourplease/domain/bid/controller/BidController.java
@@ -2,10 +2,9 @@ package com.deal4u.fourplease.domain.bid.controller;
 
 import com.deal4u.fourplease.domain.bid.dto.BidRequest;
 import com.deal4u.fourplease.domain.bid.dto.BidResponse;
-import com.deal4u.fourplease.domain.bid.entity.PageResponse;
+import com.deal4u.fourplease.domain.common.PageResponse;
 import com.deal4u.fourplease.domain.bid.service.BidService;
 import jakarta.validation.Valid;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
