+++ src/main/java/com/deal4u/fourplease/domain/bid/entity/Bid.java
@@ -1,6 +1,6 @@
 package com.deal4u.fourplease.domain.bid.entity;
 
-import com.deal4u.fourplease.domain.BaseDateEntity;
+import com.deal4u.fourplease.domain.common.BaseDateEntity;
 import com.deal4u.fourplease.domain.auction.entity.Auction;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
