+++ src/main/java/com/deal4u/fourplease/domain/file/entity/File.java
@@ -1,6 +1,6 @@
 package com.deal4u.fourplease.domain.file.entity;
 
-import com.deal4u.fourplease.domain.BaseDateEntity;
+import com.deal4u.fourplease.domain.common.BaseDateEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
