+++ src/main/java/com/deal4u/fourplease/domain/wishlist/entity/Wishlist.java
@@ -1,6 +1,6 @@
 package com.deal4u.fourplease.domain.wishlist.entity;
 
-import com.deal4u.fourplease.domain.BaseDateEntity;
+import com.deal4u.fourplease.domain.common.BaseDateEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
